<> img with avatar, + side gradients, - shit code
